# backend/docs/swagger.yaml
basePath: /
definitions:
...
  db.EducationMaterial:
    properties:
      createdAt:
        type: string
      creatorID:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      videoUrl:
        type: string
    type: object
...
  handler.CreateEducationMaterialRequest:
    properties:
      description:
        type: string
      title:
        type: string
      video_url:
        type: string
    required:
    - title
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  handler.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
...
  handler.UpdateEducationMaterialRequest:
    properties:
      description:
        type: string
      title:
        type: string
      video_url:
        type: string
    required:
    - title
    type: object
...
info:
  contact: {}
  description: Backend REST API for Centing (Stunting Prevention & Monitoring System).
  title: Centing API
  version: "1.0"
paths:
...
  /education-materials:
    get:
      description: Retrieve paginated list of education materials
      parameters:
      - description: Limit (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.EducationMaterial'
            type: array
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List education materials
      tags:
      - Education Materials
  /education-materials/{id}:
    get:
      description: Retrieve details of an education material by ID
      parameters:
      - description: Education Material ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.EducationMaterial'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get education material by ID
      tags:
      - Education Materials
...
  /nakes/education-materials:
    post:
      consumes:
      - application/json
      description: Create a new education material with title, description, and optional
        video URL
      parameters:
      - description: Create Education Material Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateEducationMaterialRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.EducationMaterial'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create education material (Nakes only)
      tags:
      - Education Materials
  /nakes/education-materials/{id}:
    delete:
      description: Delete an education material by ID
      parameters:
      - description: Education Material ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete education material (Nakes only)
      tags:
      - Education Materials
    put:
      consumes:
      - application/json
      description: Update an existing education material
      parameters:
      - description: Education Material ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update Education Material Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateEducationMaterialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.EducationMaterial'
        "400":
          description: Validation error or invalid ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update education material (Nakes only)
      tags:
      - Education Materials
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"